requirements.txt
flask
line-bot-sdk
transformers
torch
gunicorn